Started work on gameplay. Removed redundant (int) type casts from the draw() methods of the GameObject, BlockHard, BlockSoft, and Player classes. A bomb class has been created with a basic draw() method and correct collision detection to get it started. The Player class now has a dropBomb() method to allow for the player to drop bombs. The GameKeys class now supports the dropBomb() method that exists within the Player class. GameMain's listObjects List has temporarily been made static to allow for bombs to be experimented on during runtime. The GameMain class now performs a check for if the player collides with an object within listObjects. The hit() method within GameObject has been moved down past the collision methods to better represent the order of occurrence in which its methods are typically called.

+++ src/game/BlockHard.java
@@ -30,7 +30,7 @@ public class BlockHard extends GameObject {
         /*
         // For viewing the bounding box
         g.setColor(Color.RED);
-        g.fillRect(getBounds().x, getBounds().y, (int)getBounds().width, (int)getBounds().height);
+        g.fillRect(getBounds().x, getBounds().y, getBounds().width, getBounds().height);
         */
 
         g.setColor(TILE_COLOUR);
